% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscentheme.R
\name{hello_rscentheme}
\alias{hello_rscentheme}
\title{Apply rscentheme theme}
\usage{
hello_rscentheme(name = c("rose", "iris", "silver", "woodsage"))
}
\value{
nothing.
}
\description{
Apply rscentheme theme
}
