% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscentheme.R
\name{rscentheme}
\alias{rscentheme}
\title{Install and apply rscentheme theme}
\usage{
rscentheme()
}
\arguments{
\item{name}{A character name of the theme to be installed and applied}
}
\value{
nothing
}
\description{
Install and apply rscentheme theme
}
