% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{scent_palette}
\alias{scent_palette}
\title{Complete list of palettes}
\format{
An object of class \code{character} of length 8.
}
\usage{
scent_palette
}
\description{
Use \code{\link{scent_palette}} to construct palettes of desired length.
}
\keyword{datasets}
