% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscentheme.R
\name{scented_theme}
\alias{scented_theme}
\title{Apply rscentheme theme}
\usage{
scented_theme(
  name = c("rose", "iris", "mountainwater", "woodsea", "br540", "woodsea_light",
    "blanche", "pluriel")
)
}
\value{
nothing.
}
\description{
Apply rscentheme theme
}
